requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
Pillow==10.1.0
numpy==1.26.1
pandas==2.1.2
openai==1.3.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
pydantic-settings==2.0.3
SQLAlchemy==2.0.23
aiofiles==23.2.1
tenacity==8.2.3
transformers==4.35.0
torch==2.1.0
jinja2==3.1.5 